+++ Changelog.md
@@ -1,3 +1,7 @@
+### June 2026
+
+- Add opt-in TTF font fallback for missing glyphs with `GoPdf.SetFontFallback`.
+
 ### May 2016
 
 Remove old function